% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture_em-methods.R
\name{AIC.survival_ln_mixture_em}
\alias{AIC.survival_ln_mixture_em}
\title{Return AIC of a fitted \code{survival_ln_mixture_em} object}
\usage{
\method{AIC}{survival_ln_mixture_em}(model, ...)
}
\arguments{
\item{model}{A fitted \code{survival_ln_mixture_em} object.}

\item{...}{Not used.}
}
\description{
Return AIC of a fitted \code{survival_ln_mixture_em} object
}
