% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{augment.survival_ln_mixture}
\alias{augment.survival_ln_mixture}
\title{Augment data with information from a survival_ln_mixture object}
\usage{
\method{augment}{survival_ln_mixture}(x, newdata, eval_time, ...)
}
\arguments{
\item{x}{A \code{survival_ln_mixture} object.}

\item{newdata}{A \code{base::data.frame()} or \code{tibble::tiblle()} containing all
the original predictors used to create x.}

\item{eval_time}{a vector with the times where the hazard and survival distribuition
will be evaluated.}

\item{...}{Not used.}
}
\value{
A \code{tibble::tibble()} with the original covariates and ther survvial and hazard
distributions.
}
\description{
Include information about hazard and survival distribution for each individual
in a dataset.
}
