% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture_em-methods.R
\name{BIC.survival_ln_mixture_em}
\alias{BIC.survival_ln_mixture_em}
\title{Return BIC of a fitted \code{survival_ln_mixture_em} object}
\usage{
\method{BIC}{survival_ln_mixture_em}(model, ...)
}
\arguments{
\item{model}{A fitted \code{survival_ln_mixture_em} object.}

\item{...}{Not used.}
}
\description{
Return BIC of a fitted \code{survival_ln_mixture_em} object
}
