% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_cross_entropy.R
\name{binary_cross_entropy}
\alias{binary_cross_entropy}
\title{Function used to calculate the binary cross entropy (also known as log loss), given by
H_p(q) = -1/n * sum(S * log(\hat{S}) + (1 - S) * log(1 - \hat{S})), where S is the empirical survival and \hat{S} is the fitted survival.}
\usage{
binary_cross_entropy(preds, nobs = NULL, threshold = 0.005)
}
\arguments{
\item{preds}{The \verb{$preds} object from \code{plot_fit_on_data()} applied to the model. If not this one, should be a prediction \code{tibble} with the columns \code{time}, \code{strata} (if applicable), \code{estimate}, \code{.pred_survival}, \code{n.risk} (also \code{chain}, if necessary). It's important that the quantities \code{estimate} and \code{.pred_survival} are calculated for the same \code{time} and \code{strata}. It's highly recommended to simply use the object \verb{$preds} returned from the function \code{plot_fit_on_data()}.}

\item{nobs}{The number of observations used to fit the model. Can be ignored if \code{threshold} is set to 0. To easily calculate this value, use the function \code{nobs()} applied to the model object.}

\item{threshold}{Numeric value between 0 and 1. Times with \code{n.risk} below threshold * nobs will be ignored. Default is 0.005 (0.5\%).}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \code{strata}: The stratas used to fit the model.
\item \code{mean_entropy}: Mean binary cross entropy across times.
\item \code{chain}: The chain of the Bayesian model (only if necessary).
}
}
\description{
\code{binary_cross_entropy()} is used to calculate the binary cross entropy of a predictions object, preferably the object \verb{$preds} returned from \code{plot_fit_on_data()}.
}
