% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_metrics.R
\name{fit_metrics}
\alias{fit_metrics}
\title{Function used to calculate some distance metrics between the predicted survival and the observed survival.
\code{fit_metrics()} is used to calculate distance metrics between empirical and fitted survival for a predictions object, preferably the object \verb{$preds} returned from \code{plot_fit_on_data()}.}
\usage{
fit_metrics(preds, nobs = NULL, threshold = 0.005)
}
\arguments{
\item{preds}{The \verb{$preds} object from \code{plot_fit_on_data()} applied to the model. If not this one, should be a prediction \code{tibble} with the columns \code{time}, \code{strata} (if applicable), \code{estimate}, \code{.pred_survival}, \code{n.risk} (also \code{chain}, if necessary). It's important that the quantities \code{estimate} and \code{.pred_survival} are calculated for the same \code{time} and \code{strata}. It's highly recommended to simply use the object \verb{$preds} returned from the function \code{plot_fit_on_data()}.}

\item{nobs}{The number of observations used to fit the model. Can be ignored if \code{threshold} is set to 0. To easily calculate this value, use the function \code{nobs()} applied to the model object.}

\item{threshold}{Numeric value between 0 and 1. Times with \code{n.risk} below threshold * nobs will be ignored. Default is 0.005 (0.5\%). Important because the distance metrics may be too big if calculated in intervals without sufficient observations to be estimated.}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \code{strata}: The stratas used to fit the model (if necessary).
\item \code{chain}: The chain of the Bayesian model (only if necessary).
\item \code{metric}: Which metric is being calculated.
\item \code{value}: Value for the metric.
}

For now, the following metrics are available and will be included:
\itemize{
\item \code{MSE}: Mean Squared Error (the less the better).
\item \code{MAE}: Mean Absolute Error (the less the better).
\item \verb{Hellinger Distance}: Hellinger distance, sometimes called Jeffreys distance (the less the better).
\item \verb{KS Distance}: Kolmogorov-Smirnov distance (the less the better).
\item \verb{KL Divergence}: Kullback-Leiber divergence (the less the better). Measures how much the fitted survival is different from the empirical one.
}
}
\description{
Function used to calculate some distance metrics between the predicted survival and the observed survival.
\code{fit_metrics()} is used to calculate distance metrics between empirical and fitted survival for a predictions object, preferably the object \verb{$preds} returned from \code{plot_fit_on_data()}.
}
