% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistencia-package.R
\docType{package}
\name{persistencia-package}
\alias{persistencia}
\alias{persistencia-package}
\title{persistencia: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
\keyword{internal}
