% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture_em-methods.R
\name{logLik.survival_ln_mixture_em}
\alias{logLik.survival_ln_mixture_em}
\title{Return log-likelihood of a fitted \code{survival_ln_mixture_em} object}
\usage{
\method{logLik}{survival_ln_mixture_em}(model, ...)
}
\arguments{
\item{model}{A fitted \code{survival_ln_mixture_em} object.}

\item{...}{Not used.}
}
\description{
Return log-likelihood of a fitted \code{survival_ln_mixture_em} object
}
