% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_mixture_gibbs.R
\name{lognormal_mixture_gibbs}
\alias{lognormal_mixture_gibbs}
\title{LOGNORMAL CENSORED GIBBS SCHEME}
\usage{
lognormal_mixture_gibbs(
  x,
  y,
  delta,
  intercepto = F,
  numero_iteracoes = 1000,
  numero_cadeias = 1,
  valor_inicial_beta = 0
)
}
\arguments{
\item{x}{design matrix}

\item{y}{responde vector in the original scale. all must be positive, since
the log transfomation
will be taken.}

\item{M}{number of iterations}

\item{cens}{binary vector of censored observations
(0 = censored ; 1 = observed)}
}
\value{
list with the posterior distribution of each parameter.
}
\description{
Gibbs estimation of a lognormal mixture model.
(include here model equation)
}
