% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lognormal_mixture_gibbs_cpp}
\alias{lognormal_mixture_gibbs_cpp}
\title{Gibbs sampling for a two componentes mixture model}
\usage{
lognormal_mixture_gibbs_cpp(
  y,
  x,
  delta,
  numero_iteracoes,
  valor_inicial_beta = 0
)
}
\arguments{
\item{y}{\link{nx1}}

\item{x}{\link{nxp}}

\item{delta}{\link{nx1}}
}
\description{
Gibbs sampling for a two componentes mixture model
}
