% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit_on_data_em.R
\name{plot_fit_on_data_em}
\alias{plot_fit_on_data_em}
\title{Function used to quick visualize the fitted values (survival estimate) on the data used to fit the model.}
\usage{
plot_fit_on_data_em(model, type = "survival")
}
\arguments{
\item{model}{A \code{survival_ln_mixture_em} object.}

\item{type}{A character string specifying the type of plot. The default is "survival", but can be "hazard".}
}
\description{
\code{plot_fit_on_data_em()} estimates survival/hazard for the data the model was fitted on and plots the results.
}
