% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture-predict.R
\name{predict.survival_ln_mixture}
\alias{predict.survival_ln_mixture}
\title{Predict from a \code{survival_ln_mixture}}
\usage{
\method{predict}{survival_ln_mixture}(object, new_data, type = "survival", ...)
}
\arguments{
\item{object}{A \code{survival_ln_mixture} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"time"} for the survival time.
\item \code{"survival"} for the survival probability.
\item \code{"hazard"} for the hazard.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{survival_ln_mixture}
}
\examples{
train <- mtcars[1:20, ]
test <- mtcars[21:32, -1]

# Fit
mod <- survival_ln_mixture(mpg ~ cyl + log(drat), train)

# Predict, with preprocessing
predict(mod, test)

}
