% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture-fit.R
\name{survival_ln_mixture}
\alias{survival_ln_mixture}
\alias{survival_ln_mixture.default}
\alias{survival_ln_mixture.formula}
\alias{survival_ln_mixture.recipe}
\title{Fit a \code{survival_ln_mixture}}
\usage{
survival_ln_mixture(x, ...)

\method{survival_ln_mixture}{default}(x, ...)

\method{survival_ln_mixture}{formula}(formula, data, ...)

\method{survival_ln_mixture}{recipe}(x, data, ...)
}
\arguments{
\item{...}{Not currently used, but required for extensibility.}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.
\itemize{
\item The outcome must be a \strong{survival::Surv} object.
}}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
\item The outcome must be a \strong{survival::Surv} objetc.
}}
}
\value{
A \code{survival_ln_mixture} object.
}
\description{
\code{survival_ln_mixture()} fits a model.
}
\examples{

# Formula interface
library(survival)
mod <- survival_ln_mixture(Surv(time, status == 2), lung)


# Recipes interface
library(recipes)
dados <- lung \%>\% mutate(outcome = Surv(time, status == 2))
rec <- recipe(outcome ~ sex, dados)
mod3 <- survival_ln_mixture(rec, dados)

}
