% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture_em.R
\name{survival_ln_mixture_em}
\alias{survival_ln_mixture_em}
\title{Lognormal mixture model - EM algorithm for censored data}
\usage{
survival_ln_mixture_em(y, delta, X, iter = 50, mixture_components = 2)
}
\arguments{
\item{y}{Log of the events ocurrence times.}

\item{delta}{Vector containing the observations status (0 for censored, 1 for event ocurrence)}

\item{X}{Design matrix for regression. The columns are the covariates (intercept + regression covariates) and the rows are the observations.}

\item{iter}{A positive integer specifying the number of iterations for the EM algorithm.}

\item{mixture_components}{number of mixture componentes, >= 2.}

\item{...}{Not currently used, but required for extensibility.}
}
\description{
\code{survival_ln_mixture_em()} fits a lognormal mixture model using the EM algorithm for censored data, as described in LOBO, Viviana GR; FONSECA, Thaís CO; ALVES, Mariane B. Lapse risk modeling in insurance: a Bayesian mixture approach. Annals of Actuarial Science, v. 18, n. 1, p. 126-151, 2024.
}
\note{
Categorical predictors must be converted to factors before the fit,
otherwise the predictions will fail.
}
