% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture_sem.R
\name{survival_ln_mixture_sem}
\alias{survival_ln_mixture_sem}
\title{Lognormal mixture model - Stochastic EM algorithm for censored data}
\usage{
survival_ln_mixture_sem(
  y,
  delta,
  X,
  iter = 50,
  mixture_components = 2,
  starting_seed = sample(1:2^28, 1)
)
}
\arguments{
\item{y}{Log of the events ocurrence times.}

\item{delta}{Vector containing the observations status (0 for censored, 1 for event ocurrence)}

\item{X}{Design matrix for regression. The columns are the covariates (intercept + regression covariates) and the rows are the observations.}

\item{iter}{A positive integer specifying the number of iterations for the EM algorithm.}

\item{mixture_components}{Number of mixture componentes, >= 2.}

\item{starting_seed}{Starting seed for the sampler. If not specified by the user, uses a random integer between 1 and 2^28 This way we ensure, when the user sets a seed in R, that this is passed into the C++ code.}
}
\description{
\code{survival_ln_mixture_em()} fits a lognormal mixture model using the SEM algorithm for censored data, a modification of the EM described in LOBO, Viviana GR; FONSECA, Thaís CO; ALVES, Mariane B. Lapse risk modeling in insurance: a Bayesian mixture approach. Annals of Actuarial Science, v. 18, n. 1, p. 126-151, 2024.
}
\note{
Categorical predictors must be converted to factors before the fit,
otherwise the predictions will fail.
}
