% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture-methods.R
\name{nobs.survival_ln_mixture}
\alias{nobs.survival_ln_mixture}
\title{Extract the number of observations from \code{survival_ln_mixture} fit.}
\usage{
\method{nobs}{survival_ln_mixture}(object, ...)
}
\arguments{
\item{object}{A fitted \code{survival_ln_mixture} object.}

\item{...}{Not used.}
}
\value{
A single integer.
}
\description{
Extract the number of observations used in a \code{survival_ln_mixture} fit.
}
