% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture_em-methods.R
\name{plot.survival_ln_mixture_em}
\alias{plot.survival_ln_mixture_em}
\title{Visualizes the path of the EM algorithm}
\usage{
\method{plot}{survival_ln_mixture_em}(x, ...)
}
\arguments{
\item{x}{A fitted \code{survival_ln_mixture_em} object.}

\item{...}{Not used.}
}
\value{
A \code{ggplot} object (or \code{plotly}, if the package is avaiable) of the EM algorithm iterations.
}
\description{
Visualizes the path of the EM algorithm
}
