% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ln_mixture-fit.R
\name{survival_ln_mixture}
\alias{survival_ln_mixture}
\alias{survival_ln_mixture.default}
\alias{survival_ln_mixture.formula}
\title{Lognormal mixture model}
\usage{
survival_ln_mixture(
  formula,
  data,
  intercept = TRUE,
  iter = 1000,
  warmup = floor(iter/10),
  thin = 1,
  chains = 1,
  cores = 1,
  ...
)

\method{survival_ln_mixture}{default}(formula, ...)

\method{survival_ln_mixture}{formula}(formula, data, intercept = TRUE, ...)
}
\arguments{
\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side. The outcome must be a \link[survival:Surv]{survival::Surv}
object.}

\item{data}{A \strong{data frame} containing both the predictors and the outcome.}

\item{intercept}{A logical. Should an intercept be included in the processed data?}

\item{iter}{A positive integer specifying the number of iterations for each chain (including warmup).}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin) iterations per chain.
The number of warmup iterations should be smaller than iter.}

\item{thin}{A positive integer specifying the period for saving samples.}

\item{chains}{A positive integer specifying the number of Markov chains.}

\item{cores}{Number of cores to use when executing the chains in parallel. The number of
cores used is limited by the number of avaliable cores and the number of chains. The number
of available cores can be checked with \code{parallel::detectCores(logical = FALSE)}.}

\item{...}{Not currently used, but required for extensibility.}
}
\value{
A \code{survival_ln_mixture} object, which is a list with the following componentes:

\item{posterior}{A \link[posterior:draws_matrix]{posterior::draws_matrix} with the posterior of the parameters of the model.}
\item{nobs}{A integer holding the number of observations used to generate the fit.}
\item{blueprint}{The blueprint component of the output of \link[hardhat:mold]{hardhat::mold}}
}
\description{
\code{survival_ln_mixture()} fits a lognormal mixture model, as described in (referencia artigo viviana).
Colocar mais detalhes sobre o modelo (como a equacao) em details.
}
\note{
Categorical predictos must be converted to factores before the fit,
otherwise the predictions will fail.
}
\examples{

# Formula interface
library(survival)
set.seed(1)
mod <- survival_ln_mixture(Surv(time, status == 2) ~ NULL, lung, intercept = TRUE)

}
